divide primary crate into more sub crates
