+++ crates/plotinator-file-io/src/dropped_files/preview_dropped.rs
