+++ crates/plotinator-file-io/src/file_dialog/native.rs
@@ -1,18 +1,17 @@
 use std::{fs, path::PathBuf};
 
-use crate::{
-    app::{
-        custom_files::{
-            CUSTOM_HEADER_PLOT_DATA, CUSTOM_HEADER_PLOT_UI_STATE, CustomFileContent,
-            try_parse_custom_file,
-        },
-        file_dialog::{FILE_FILTER_EXTENSIONS, FILE_FILTER_NAME},
-        loaded_files::LoadedFiles,
-    },
-    plot::LogPlotUi,
-};
+use plotinator_plot_ui::LogPlotUi;
 use plotinator_supported_formats::SupportedFormat;
 use serde::Serialize;
+
+use crate::{
+    custom_files::{
+        CUSTOM_HEADER_PLOT_DATA, CUSTOM_HEADER_PLOT_UI_STATE, CustomFileContent,
+        try_parse_custom_file,
+    },
+    file_dialog::{FILE_FILTER_EXTENSIONS, FILE_FILTER_NAME},
+    loaded_files::LoadedFiles,
+};
 
 #[derive(Debug, Default)]
 pub struct NativeFileDialog {
@@ -21,7 +20,7 @@ pub struct NativeFileDialog {
 
 impl NativeFileDialog {
     /// Opens a native file dialog to pick multiple files.
-    pub(crate) fn open(&mut self) {
+    pub fn open(&mut self) {
         if let Some(paths) = rfd::FileDialog::new()
             .add_filter(FILE_FILTER_NAME, FILE_FILTER_EXTENSIONS)
             .pick_files()
@@ -31,7 +30,7 @@ impl NativeFileDialog {
     }
 
     /// Saves the plot UI state to a file.
-    pub(crate) fn save_plot_ui(plot_ui: &LogPlotUi) {
+    pub fn save_plot_ui(plot_ui: &LogPlotUi) {
         Self::save_data_to_file(
             plot_ui,
             "Save Plot UI State",
@@ -41,7 +40,7 @@ impl NativeFileDialog {
     }
 
     /// Saves the plot data to a file.
-    pub(crate) fn save_plot_data(
+    pub fn save_plot_data(
         plot_files: &[SupportedFormat],
         #[cfg(all(not(target_arch = "wasm32"), feature = "mqtt"))] mqtt_plots: Option<
             &plotinator_mqtt::MqttPlotData,
@@ -106,7 +105,7 @@ impl NativeFileDialog {
 
     /// Parses all picked files and loads them into the application.
     /// Returns an `Option<LogPlotUi>` if a plot UI state file was loaded.
-    pub(crate) fn parse_picked_files(
+    pub fn parse_picked_files(
         &mut self,
         loaded_files: &mut LoadedFiles,
     ) -> anyhow::Result<Option<Box<LogPlotUi>>> {
