+++ crates/plotinator-mqtt-ui/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "plotinator-mqtt-ui"
+authors.workspace = true
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+
+[dependencies]
+egui.workspace = true
+plotinator-mqtt.workspace = true
+
+[lints]
+workspace = true
