+++ crates/plotinator-plot-ui/Cargo.toml
@@ -1,0 +1,45 @@
+[package]
+name = "plotinator-plot-ui"
+authors.workspace = true
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+
+[dependencies]
+plotinator-supported-formats.workspace = true
+plotinator-macros.workspace = true
+plotinator-proc-macros.workspace = true
+plotinator-plot-util.workspace = true
+plotinator-ui-util.workspace = true
+plotinator-strfmt.workspace = true
+plotinator-log-if.workspace = true
+plotinator-mqtt = { workspace = true, optional = true }
+plotinator-mqtt-ui = { workspace = true, optional = true }
+egui.workspace = true
+egui_plot.workspace = true
+egui-phosphor.workspace = true
+egui-notify.workspace = true
+chrono.workspace = true
+log.workspace = true
+serde.workspace = true
+strum.workspace = true
+strum_macros.workspace = true
+smallvec.workspace = true
+rayon.workspace = true
+
+profiling = { workspace = true, optional = true }
+puffin = { workspace = true, optional = true }
+puffin_http = { workspace = true, optional = true }
+
+[features]
+default = ["mqtt"]
+mqtt = ["dep:plotinator-mqtt", "dep:plotinator-mqtt-ui"]
+log_time = ["plotinator-proc-macros/log_time"]
+profiling = ["dep:profiling", "dep:puffin", "dep:puffin_http", "plotinator-proc-macros/profiling"]
+
+
+[lints]
+workspace = true
