+++ crates/plotinator-plot-ui/src/click_delta.rs
