+++ crates/plotinator-plot-ui/src/plot_settings/log_groups.rs
