+++ crates/plotinator-plot-ui/src/plot_settings/mipmap_settings.rs
