+++ crates/plotinator-plot-ui/src/plot_settings/plot_visibility_config.rs
