+++ crates/plotinator-plot-ui/src/plot_settings/series_plot_settings.rs
