+++ crates/plotinator-plot-ui/src/x_axis_formatter.rs
